+++ Assets/_Scripts/System/Mining/MiningSkillTree.cs
@@ -49,6 +49,10 @@ public class MiningSkillTree : MonoBehaviour
     [SerializeField] private Text CostText;
     [SerializeField] private Button BuyButton;
 
+
+    private int _id = -1;
+
+
     private void UpdateUI()
     {
         SkillPointsText.text = "Skill Points: " + UISystem.Instance.NumberFormatInt(skillPoints);
@@ -70,11 +74,13 @@ public class MiningSkillTree : MonoBehaviour
             SkillPointsButton.GetComponent<Button>().interactable = false;
             SkillPointsButton.GetComponent<Button>().onClick.RemoveAllListeners();
         }
+
+        if (_id != -1)
+            ShowSkillData(_id);
     }
 
     private void AddSkillPoint()
     {
-        Debug.Log($"Add skill point to {requiredMiningLevel} | {MiningSystem.Instance.MiningLevel}");
         if(MiningSystem.Instance.MiningLevel >= requiredMiningLevel)
         {
             MiningSystem.Instance.RemoveMiningLevel(requiredMiningLevel);
@@ -93,17 +99,17 @@ public class MiningSkillTree : MonoBehaviour
         BonusText.text = skills[id].name + " (" + skills[id].level + "/" + skills[id].maxLevel + ")";
         CostText.text = $"Cost: {UISystem.Instance.NumberFormat(skills[id].cost)}";
 
-        if(skillPoints <= 0)
-        {
-            BuyButton.GetComponent<Image>().color = UISystem.Instance.buyButtonDisabledColor;
-            BuyButton.interactable = false;
-            BuyButton.GetComponentInChildren<Text>().text = "No Skill Points";
-        }
-        else if(skills[id].level >= skills[id].maxLevel)
+        if(skills[id].level >= skills[id].maxLevel)
         {
             BuyButton.GetComponent<Image>().color = UISystem.Instance.buyButtonMaxedColor;
             BuyButton.interactable = false;
             BuyButton.GetComponentInChildren<Text>().text = "Maxed";
+        }
+        else if(skillPoints < skills[id].cost)
+        {
+            BuyButton.GetComponent<Image>().color = UISystem.Instance.buyButtonDisabledColor;
+            BuyButton.interactable = false;
+            BuyButton.GetComponentInChildren<Text>().text = "No Skill Points";
         }
         else
         {
@@ -115,6 +121,7 @@ public class MiningSkillTree : MonoBehaviour
         BuyButton.onClick.RemoveAllListeners();
         BuyButton.onClick.AddListener(() => BuySkill(id));
 
+        _id = id;
     }
 
     public void BuySkill(int id)
@@ -131,25 +138,24 @@ public class MiningSkillTree : MonoBehaviour
         if(skills[id].level != 0)
             skills[id].isPurchased = true;
 
+        _id = id;
         AddBonus(id);
         UpdateUI();
         ShowSkillData(id);
         SaveSkill(id);
     }
 
-    //TODO
     private void AddBonus(int id)
     {
         switch (id)
         {
             case 0:
-                Debug.Log("Add bonus to skill 0");
+                MiningSystem.Instance.SetAutoMining(true);
                 break;
             default:
                 break;
         }
     }
-
 
     private void SaveSkillPoints()
     {
@@ -165,7 +171,10 @@ public class MiningSkillTree : MonoBehaviour
     {
         skills[id].level = PlayerPrefs.GetInt("MiningSkillTree_" + id, 0);
         if (skills[id].level != 0)
+        {
             skills[id].isPurchased = true;
+            AddBonus(id);
+        }
     }
 
     private void Start()